Update and streamline build_vanagon
This does a couple things. 1) It renames non-arm to x86, since that's the only arch we're building with that matrix. 2) It disables the qemu-user-static installation, because we only need that if we're building for a different arch than the host platform, and we removed ppc64le. 3) It sets the x86 Ruby version to 3.2.7 to match arm. 4) It removes caching the bundle across workflow runs. I had only had bundler-cache set to true to get it to bundle install automatically, but it also tries caching this, which not only isn't really needed and causes warnings in the workflow due to cache name collisions, but also isn't really wanted, because we want to start fresh each time.