Validate Semantic Version syntax on release preparation
The release preparation task in liberal in the input it accept and pass
it directly to the underlying (project specific) to bump the version.

These underlying projects seems to also be liberal in what they accept,
and we already ended-up releasing pre-release gems with an invalid
version number according to SemVer (<major>.<minor>.<path>.<pre-release>
instead of <major>.<minor>.<patch>-<pre-release>).

I feel like ruby does some magic with version numbers (i.e.
`Gem::Version.new('5.0.0-rc1').to_s #=> "5.0.0.pre.rc1"`), so it is
probably not a big deal, but the tag name and the version in the PR
should really follow semver as we tell our consumers that it is what we
use.