Add a reusable workflow for Beaker acceptance
This workflow was extracted from the common elements of openvox,
openvox-agent and openvox-server .github/workflows/acceptance.{yaml,yml}.

It's intended to replace the guts of each of those and to be useable as
a callable acceptance workflow from future workflows for openvoxdb and
any other openvox project we need to run beaker on some set of vms for.

Or this could be called within pr testing workflows that first call the
build workflows.