Do not require fork param in beaker_acceptance.yml
I made a mistake when I added this parameter in that I gave it a default
and set required to true. Consequently a calling workflow is forced
to set fork even if it would just want the default...

Setting required to false so that callers don't have to pass a fork
value unless they want to.