Handle install flags when not passed
In a callable workflow a boolean input that isn't given will be false,
making it impossible to differentiate between the caller didn't pass a
value (so we should supply a default), versus the caller explicitly
passed false and we should respect that.

To work around this, switched install-openvox-server and
install-openvoxdb to type string, so we can handle them, and then use
fromJson to get the boolean value again from the outputs used by the rest
of the workflow.