Add Ubuntu 25.04, Debian 13 to ezbake builds and use ezbake-ver
Ubuntu 25.04 is new and we were missing Debian 13 here. This also wires up the ezbake-ver variable when given. We do it this way because the build rake task looks for the env var to exist to decide if it passes it through to the lein invocation, and we don't want to set it at all if we aren't passing in a value.